# Move OCSPRequest implementation to Rust