--- a/src/cryptography/hazmat/backends/openssl/backend.py
+++ b/src/cryptography/hazmat/backends/openssl/backend.py
@@ -420,12 +420,6 @@ class Backend(BackendInterface):
             get_ext=self._lib.X509_CRL_get_ext,
             handlers=_CRL_EXTENSION_HANDLERS,
         )
-        self._ocsp_req_ext_parser = _X509ExtensionParser(
-            self,
-            ext_count=self._lib.OCSP_REQUEST_get_ext_count,
-            get_ext=self._lib.OCSP_REQUEST_get_ext,
-            rust_callback=asn1.parse_ocsp_req_extension,
-        )
         self._ocsp_basicresp_ext_parser = _X509ExtensionParser(
             self,
             ext_count=self._lib.OCSP_BASICRESP_get_ext_count,