--- a/src/cryptography/hazmat/backends/openssl/backend.py
+++ b/src/cryptography/hazmat/backends/openssl/backend.py
@@ -85,8 +85,7 @@ from cryptography.hazmat.backends.openssl.x509 import (
     _CertificateSigningRequest,
     _RevokedCertificate,
 )
-from cryptography.hazmat.bindings._rust import asn1
-from cryptography.hazmat.bindings._rust.ocsp import parse_ocsp_resp_extension
+from cryptography.hazmat.bindings._rust import asn1, ocsp as rust_ocsp
 from cryptography.hazmat.bindings.openssl import binding
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import (
@@ -422,7 +421,7 @@ class Backend(BackendInterface):
             self,
             ext_count=self._lib.OCSP_BASICRESP_get_ext_count,
             get_ext=self._lib.OCSP_BASICRESP_get_ext,
-            rust_callback=parse_ocsp_resp_extension,
+            rust_callback=rust_ocsp.parse_ocsp_resp_extension,
         )
         self._ocsp_singleresp_ext_parser = _X509ExtensionParser(
             self,