--- a/src/cryptography/hazmat/bindings/_rust/ocsp.pyi
+++ b/src/cryptography/hazmat/bindings/_rust/ocsp.pyi
@@ -1,0 +1,3 @@
+from cryptography.x509.ocsp import OCSPRequest
+
+def load_der_ocsp_request(data: bytes) -> OCSPRequest: ...