--- a/src/cryptography/hazmat/bindings/_rust/ocsp.pyi
+++ b/src/cryptography/hazmat/bindings/_rust/ocsp.pyi
@@ -1,3 +1,7 @@
+from cryptography.x509 import ExtensionType
 from cryptography.x509.ocsp import OCSPRequest
 
 def load_der_ocsp_request(data: bytes) -> OCSPRequest: ...
+def parse_ocsp_resp_extension(
+    der_oid: bytes, ext_data: bytes
+) -> ExtensionType: ...