Has the core structure of the parser, but lots of edge cases missed. Also opportunities for code-cleanup still present.