@alex 2021-05-22
Performance numbers:

# Before (3.4.7)
```
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
50000 loops, best of 5: 6.92 usec per loop
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
50000 loops, best of 5: 7.25 usec per loop
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
50000 loops, best of 5: 7.32 usec per loop
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
50000 loops, best of 5: 7.11 usec per loop
```

# After (this branch)
```
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
500000 loops, best of 5: 596 nsec per loop
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
500000 loops, best of 5: 584 nsec per loop
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
500000 loops, best of 5: 611 nsec per loop
(.venv) alex@media-01:~/t$ python3 -mtimeit -s "from cryptography.x509.ocsp import load_der_ocsp_request; data = open('req-ext-unknown-oid.der', 'rb').read()" "load_der_ocsp_request(data)"
500000 loops, best of 5: 589 nsec per loop
```